# internal/docs/swagger.yaml
basePath: /api
definitions:
  handler.BindRepoRequest:
    properties:
      path:
        type: string
      remotes:
        type: string
    required:
    - path
    type: object
  handler.CloneRepoRequest:
    properties:
      path:
        type: string
      url:
        type: string
    required:
    - path
    - url
    type: object
...
  handler.GitActionRequest:
    properties:
      files:
        items:
          type: string
        type: array
      id:
        type: string
      message:
        description: For commit
        type: string
    required:
    - id
    type: object
...
  handler.NewRepoRequest:
    properties:
      path:
        type: string
    required:
    - path
    type: object
...
  handler.UnbindRepoRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
host: localhost:1984
info:
  contact: {}
  description: VibeGo 后端服务 API
  title: VibeGo API
  version: 0.0.1
paths:
...
  /api/git/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Files to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.GitActionRequest'
      produces:
      - application/json
      responses: {}
      summary: Stage files
      tags:
      - Git
  /api/git/bind:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bind repo request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BindRepoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Bind git repository
      tags:
      - Git
  /api/git/checkout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Files to restore
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.GitActionRequest'
      produces:
      - application/json
      responses: {}
      summary: Discard changes (Checkout file from index/HEAD)
      tags:
      - Git
  /api/git/clone:
    post:
      consumes:
      - application/json
      parameters:
      - description: Clone repo request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CloneRepoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Clone git repository
      tags:
      - Git
  /api/git/commit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Commit message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.GitActionRequest'
      produces:
      - application/json
      responses: {}
      summary: Commit staged changes
      tags:
      - Git
  /api/git/diff:
    get:
      parameters:
      - description: Repo ID
        in: query
        name: id
        required: true
        type: string
      - description: File path
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get file diff
      tags:
      - Git
...
  /api/git/log:
    get:
      parameters:
      - description: Repo ID
        in: query
        name: id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get git commit log
      tags:
      - Git
  /api/git/new:
    post:
      consumes:
      - application/json
      parameters:
      - description: New repo request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.NewRepoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Initialize new git repository
      tags:
      - Git
  /api/git/reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: Files to reset
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.GitActionRequest'
      produces:
      - application/json
      responses: {}
      summary: Unstage files (Reset)
      tags:
      - Git
  /api/git/show:
    get:
      parameters:
      - description: Repo ID
        in: query
        name: id
        required: true
        type: string
      - description: File path
        in: query
        name: path
        required: true
        type: string
      - description: 'Ref (default: HEAD)'
        in: query
        name: ref
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Show file content at specific ref
      tags:
      - Git
  /api/git/status:
    get:
      parameters:
      - description: Repo ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get git status
      tags:
      - Git
  /api/git/unbind:
    post:
      consumes:
      - application/json
      parameters:
      - description: Unbind request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UnbindRepoRequest'
      produces:
      - application/json
      responses: {}
      summary: Unbind repo
      tags:
      - Git
  /api/git/undo_commit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Repo ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.GitActionRequest'
      produces:
      - application/json
      responses: {}
      summary: Undo last commit (Soft reset to HEAD~1)
      tags:
      - Git
...
swagger: "2.0"